Parse Tweet Dataset
The Tweet class represents a single Tweet.
The locations will need to be intelligently parsed.
For now, the creation time is a String because Datetimes are annoying

The TweetParser takes an InputStream and turns it into a List of
  Tweets based on the ~ separation character in our current dataset
Ideally, it will be extended to work directly with the Twitter
  Stream API.

Currently Main takes the dataset on standard input and parses it.
It doesn't do much with it.
Next steps are to feed it into Elasticsearch.