Subclass JSONOutputter to get useful things
This depends on a very lightly modified version of the Stanford
CoreNLP library.

Options which are not set by the Pipeline are now explicitly set
because the default values are not what we want.

This significantly trims down the size of each JSON document. It will
also allow us to easily include our custom annotations in our JSON
documents.